Move WithClusterSuffix to utils package
